+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- `get_boundary_condition` at class `Parameters`. It returns the appropriate boundary parameters that are
+expected by the derivative functions. by [@fschuch](https://github.com/fschuch).
+
 ## [0.1.8] - 2020-09-29
 
 ### Fixed